+++ __init__.py
@@ -2,3 +2,5 @@ from version import version as __version__
 from .regression import *
 from .ordination import *
 from .diversity import *
+from .matrix_comp import *
+from utils import *