+++ matrix_comp/__init__.py
@@ -1,0 +1,5 @@
+from mantel import Mantel
+from anosim import anosim
+from simper import simper
+from procrust_test import procrustes_test
+from fourthcorner import corner4